FIX integrate where parameter comes from combat, supply
- combat, supply Page 중 어디에서 넘어온 변수인지 확인하는 작업
- getCurrentPage는 `combat` parameter를 기준으로 버튼을 클릭 했을 때 전체 `value`값이 0이면 현재 `supply`에서 테스트 중인것을 알 수 있다. (기본적으로 모든 문제에 배점이 1점이상 주어졌기 때문)
- getCurrentPage에서 확인된 파라미터가 getMaxValueKey로 넘어가서  maxKey값을 return 해준다.
- resultPage로 변수를 넘겨 줄 수 있다.
